Added ranking commands

+++ Operator/operator.js
@@ -9,12 +9,16 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
     });
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+//‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾\
+// |‾‾‾‾  |‾‾‾ ‾‾|‾‾ |   | |‾‾| | 
+// └────┐ ├──    |   |   | |──┘ |
+//  ____| |___   |   |___| |    |
+//______________________________/
+const discord_js_1 = require("discord.js");
 const child_process_1 = require("child_process");
 const fs = require('fs');
 const can = require('canvas');
-const { Client, Intents, Message } = require('discord.js');
-const { MessageActionRow, MessageButton, MessageEmbed, MessageAttachment, WebhookClient } = require('discord.js');
-const client = new Client({ partials: [Partials.Message, Partials.Channel, Partials.Reaction, Partials.GuildMember, Partials.User], intents: 131071 });
+const client = new discord_js_1.Client({ partials: [discord_js_1.Partials.Message, discord_js_1.Partials.Channel, discord_js_1.Partials.Reaction, discord_js_1.Partials.GuildMember, discord_js_1.Partials.User], intents: 131071 });
 let server;
 //‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾‾\
 //  |‾‾‾‾ |    | |\  | |‾‾‾ ‾‾|‾‾ ‾‾|‾‾  |‾‾‾| |\  | |‾‾‾‾  |
@@ -45,49 +49,51 @@ client.on('ready', () => __awaiter(void 0, void 0, void 0, function* () {
     }
 }));
 client.on('interactionCreate', (interaction) => __awaiter(void 0, void 0, void 0, function* () {
-    if (interaction.commandName == 'start' && checkOwner(interaction)) {
-        if (server == undefined || server.killed) {
-            yield interaction.reply('Operator has started the development bot.');
+    if (interaction instanceof discord_js_1.CommandInteraction) {
+        if (interaction.commandName == 'start' && checkOwner(interaction)) {
+            if (server == undefined || server.killed) {
+                yield interaction.reply('Operator has started the development bot.');
+                server = (0, child_process_1.fork)('./index.js');
+                server.on('message', function (data) {
+                    console.log(data.toString());
+                });
+                server.on('error', function (error) {
+                    return __awaiter(this, void 0, void 0, function* () {
+                        console.log('Operator instance crashed with ' + error.message);
+                        if (error != null) {
+                            yield interaction.followUp(`Operator detected a bot crash.\n**Error Code:** ${error.message}`);
+                        }
+                    });
+                });
+            }
+        }
+        else if (interaction.commandName == 'stop' && checkOwner(interaction) && server) {
+            if (!server.killed) {
+                yield interaction.reply('Closing current operator instance.');
+                if (server.kill()) {
+                    yield interaction.followUp('Operator has closed the bot successfully.');
+                }
+                else {
+                    yield interaction.followUp('Operator was unable to properly close the current instance.');
+                }
+            }
+        }
+        else if (interaction.commandName == 'restart' && checkOwner(interaction) && server && !server.killed) {
+            yield interaction.reply('Operator has restarted the development bot.');
+            server.kill();
             server = (0, child_process_1.fork)('./index.js');
             server.on('message', function (data) {
                 console.log(data.toString());
             });
-            server.on('close', function (code) {
+            server.on('error', function (code) {
                 return __awaiter(this, void 0, void 0, function* () {
-                    console.log('Operator instance crashed with ' + code);
+                    console.log('Operator instance crashed with ' + code.message);
                     if (code != null) {
-                        yield interaction.followUp(`Operator detected a bot crash.\n**Error Code:** ${code}`);
+                        yield interaction.followUp(`Operator detected a bot crash.\n**Error Code:** ${code.message}`);
                     }
                 });
             });
         }
-    }
-    else if (interaction.commandName == 'stop' && checkOwner(interaction) && server) {
-        if (!server.killed) {
-            yield interaction.reply('Closing current operator instance.');
-            if (server.kill()) {
-                yield interaction.followUp('Operator has closed the bot successfully.');
-            }
-            else {
-                yield interaction.followUp('Operator was unable to properly close the current instance.');
-            }
-        }
-    }
-    else if (interaction.commandName == 'restart' && checkOwner(interaction) && server && !server.killed) {
-        yield interaction.reply('Operator has restarted the development bot.');
-        server.kill();
-        server = (0, child_process_1.fork)('./index.js');
-        server.on('message', function (data) {
-            console.log(data.toString());
-        });
-        server.on('close', function (code) {
-            return __awaiter(this, void 0, void 0, function* () {
-                console.log('Operator instance crashed with ' + code);
-                if (code != null) {
-                    yield interaction.followUp(`Operator detected a bot crash.\n**Error Code:** ${code}`);
-                }
-            });
-        });
     }
 }));
 client.login(require('./token.json').token);
